Reorganized MM package structure to fit CA
