+++ matchmaker/src/main/java/dev/totallyspies/spydle/matchmaker/use_case/AutoscaleService.java
@@ -1,4 +1,4 @@
-package dev.totallyspies.spydle.matchmaker.service;
+package dev.totallyspies.spydle.matchmaker.use_case;
 
 import dev.totallyspies.spydle.matchmaker.generated.model.AutoscaleRequestModelRequestStatus;
 import org.slf4j.Logger;
@@ -6,9 +6,9 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 @Service
-public class CustomFleetAutoscalerService {
+public class AutoscaleService {
 
-    private final Logger logger = LoggerFactory.getLogger(CustomFleetAutoscalerService.class);
+    private final Logger logger = LoggerFactory.getLogger(AutoscaleService.class);
 
     public int autoscale(AutoscaleRequestModelRequestStatus currentState) {
         // TODO configure based on env
